requirements.txt
aiogram==3.4.1
gspread==5.12.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
python-dotenv==1.0.1
pytest==7.4.3
pytest-asyncio==0.21.1